requirements.txt
streamlit
pypdf
PyPDF2
pymupdf
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
pysqlite3-binary
